Update SendGrid service to use MailService object
The import statement at the top of the email-service file has been changed to import the entire MailService from the Sendgrid. This change allows the setApiKey and send methods to be called directly on the MailService, increasing consistency and readability. The usage of Sendgrid services in the application has been updated accordingly.